[WIP] Convert to pip module structure
