--- data/DFAST_VEGAS_2_DK.csv
@@ -1,1 +1,0 @@
-﻿
