+++ draftfast/nfl/data_cleaning_constants.py
