+++ draftfast/scripts/prepare_contest_data.sh
