+++ draftfast/scripts/prepare_nfl_contest_data.sh
