+++ setup.py
@@ -1,0 +1,21 @@
+import setuptools
+
+with open('README.md', 'r') as fh:
+    long_description = fh.read()
+
+setuptools.setup(
+    name='draftfast',
+    version='0.0.1',
+    author='Ben Brostoff',
+    author_email='ben.brostoff@gmail.com',
+    description='A tool to automate and optimize DraftKings and FanDuel lineup construction.',
+    long_description=long_description,
+    long_description_content_type='text/markdown',
+    url='https://github.com/BenBrostoff/draft-kings-fun',
+    packages=setuptools.find_packages(),
+    classifiers=[
+        'Programming Language :: Python :: 3',
+        'License :: OSI Approved :: MIT License',
+        'Operating System :: OS Independent',
+    ],
+)
